requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
tenacity
python-dotenv
